+++ ltv/src/writer.rs
@@ -3,7 +3,7 @@ use crate::ByteOrder;
 use std::io;
 
 pub trait LTVContainer<const ED: ByteOrder, const LENGTH_SIZE: usize> {
-    fn write_ltv<'a, T>(&mut self, obj_id: usize, obj: &T) -> io::Result<usize>
+    fn write_ltv<'a, T>(&mut self, obj_id: u8, obj: &T) -> io::Result<usize>
     where
         T: LTVItem<ED>;
 }
@@ -11,7 +11,7 @@ pub trait LTVContainer<const ED: ByteOrder, const LENGTH_SIZE: usize> {
 impl<W: io::Write, const ED: ByteOrder, const LENGTH_SIZE: usize> LTVContainer<ED, LENGTH_SIZE>
     for W
 {
-    fn write_ltv<'a, T: LTVItem<ED>>(&mut self, obj_id: usize, obj: &T) -> io::Result<usize> {
+    fn write_ltv<'a, T: LTVItem<ED>>(&mut self, obj_id: u8, obj: &T) -> io::Result<usize> {
         let data = obj.to_ltv();
         if data.len() == 0 {
             return Ok(0);
@@ -58,7 +58,7 @@ impl<W: LTVContainer<ED, LENGTH_SIZE>, const ED: ByteOrder, const LENGTH_SIZE: u
 impl<W: LTVContainer<ED, LENGTH_SIZE>, const ED: ByteOrder, const LENGTH_SIZE: usize>
     LTVContainer<ED, LENGTH_SIZE> for LTVWriter<W, ED, LENGTH_SIZE>
 {
-    fn write_ltv<'a, T: LTVItem<ED>>(&mut self, obj_id: usize, obj: &T) -> io::Result<usize> {
+    fn write_ltv<'a, T: LTVItem<ED>>(&mut self, obj_id: u8, obj: &T) -> io::Result<usize> {
         self.writer.write_ltv(obj_id, obj)
     }
 }
